requirements.txt
allure-pytest==2.14.0
Faker==37.1.0
pytest==8.3.5
python-dotenv==1.0.0
requests==2.32.3